move ffi wrappers next to their related code
